Add files via upload
Explaining hypothesis 
 
The analysis aimed to observe the relationship between review, the sentiment of helpfulness and Rating of products. It is believed that in this way it is possible to understand how the public feels about products and brands. The biggest observation was that the number of words significantly impacts the product rating and the sentiment of help that the review can bring to the potential customer, factors that we assume can contribute to the position of a product in Sales Rank, once reviews are often a key point on customer decision to buy. For clothing and shoes, reviews will inform if it is true to size if the customer that usually buys M should buy one measure up or down if the material is as it looks on the picture etc…  
This analysis also helps us understand which titles and brands are most reviewed. Clotting being the biggest category reviewed and with titles in better positions in Sales Rank as said before leads us to conclude that customers who buy clothes online need reviews to make this decision. 
In conclusion, the dataset has a large number of positive reviews, and these products usually have a high rate of helpful reviews which rank better within the sales rank. The customer needs to be stimulated to write more words, useful words, which helps the public's decision to buy that product or not.